pyproject.toml
[tool.poetry]
name = "scrapi"
version = "0.1.0"
description = ""
authors = ["hdeep03 <hdeep03@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
playwright = "^1.47.0"
fastapi = "^0.115.0"
requests = "^2.32.3"
pypdf = "^5.0.1"
pydantic = "^2.9.2"
pytest = "^8.3.3"
groq = "^0.11.0"
instructor = "^1.5.1"
google-generativeai = "^0.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
